delete search page, changed successful verify
