+++ .gitignore
@@ -1,4 +1,4 @@
 /venv/
 /.pytest_cache/
 .idea
-/test/test_start_page.py
+
